немного отрефакторил код нашего приложения. добавил файл Extensions.kt. В Extensions.kt написал два метода которые в дальнейшем возможно применю. переместил список гордов в файл City.kt.
