+++ app/src/main/java/com/bignerdranch/android/androidwithkotlin/Extensions.kt
@@ -1,0 +1,25 @@
+package com.bignerdranch.android.androidwithkotlin
+
+import android.content.Context
+import android.view.View
+import android.view.inputmethod.InputMethodManager
+import android.widget.EditText
+
+fun EditText.showKeyboard() {
+    val imm = context.getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager
+    this.requestFocus()
+    imm.showSoftInput(this, 0)
+}
+
+
+fun View.hideKeyboard(): Boolean {
+    try {
+        val inputMethodManager =
+            context.getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager
+        return inputMethodManager.hideSoftInputFromWindow(windowToken, 0)
+    } catch (ignored: RuntimeException) {
+    }
+    return false
+}
+
+//пока что оставлю этот файл . В дальнейшем эти методы использую.
