+++ app/src/main/java/com/bignerdranch/android/androidwithkotlin/model/repository/RepositoryImpl.kt
@@ -1,14 +1,13 @@
 package com.bignerdranch.android.androidwithkotlin.model.repository
 
+import com.bignerdranch.android.androidwithkotlin.model.entities.City
 import com.bignerdranch.android.androidwithkotlin.model.entities.Weather
-import com.bignerdranch.android.androidwithkotlin.model.entities.getRussianCities
-import com.bignerdranch.android.androidwithkotlin.model.entities.getWorldCities
 
 class RepositoryImpl : Repository {
     override fun getWeatherFromServer() = Weather()
-    override fun getWeatherFromLocalStorageRus() = getRussianCities()
+    override fun getWeatherFromLocalStorageRus() = City.getRussianCities()
 
-    override fun getWeatherFromLocalStorageWorld() = getWorldCities()
+    override fun getWeatherFromLocalStorageWorld() = City.getWorldCities()
 
 
 }
